Here is a new project from Richard Cavell.
It's an assembly language program for the 6809 CPU.  I am developing it with
the TRS-80 Color Computer 2B in mind.

This program measures the length of a C-style string, and returns the 16-bit
result.

It includes a test helper utility. A BASIC program is to follow.

Project started July 2017.

I hope you like it!

Signed-off-by: Richard Cavell <richardcavell@mail.com>