Release of version 1.4 of 6809-strlen!
    * Addition of two new files :
      * Beginner_exercises.txt
      * Advanced_exercises.txt

    * Update the version numbers and dates in each file that has them

This project is now complete.  I will continue to contemplate it from time to
time, and I will maintain it as necessary, so it is still an active project.

But I see no reason for anything to be added, removed or modified for the
foreseeable future.  It has met its objectives.

Thanks for reading.  Goodbye now.  Any questions, please contact me.

Signed-off-by: Richard Cavell <richardcavell@mail.com>